fix: Remove outline-view display, flex-drection, height, width:
Using `display: flex;` and `flex-direction: column;` do not make sense
when it has a unique element as a child (the `ul`), so removing those
rules does not breaks the component's styles, and as Atom's styles set
the panes to an absolute position and `top`, `right`, `bottom` and
`left` to `0` it takes all the width and height. That let's remove the
`width` and `height` rules.

Co-Authored-By: Alejandro Bar Acedo <abaracedo@users.noreply.github.com>